domingo 05 de nayo

+++ database/migrations/2024_04_12_210720_create_category_lesson_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('category_lesson', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('category_id');
-            $table->unsignedInteger('lesson_id');
+            $table->unsignedBigInteger('category_id');
+            $table->unsignedBigInteger('lesson_id');
 
             $table->foreign('category_id')->references('id')->on('categories')->onDelete('cascade');
             $table->foreign('lesson_id')->references('id')->on('lessons')->onDelete('cascade');
